Refactor namespace controller to use workers, do more delete collection calls
